Fix CODE ENDS problem and add new feature
if the tag <CODE ENDS> is in the document
within the last '}' like "} <CODE ENDS>"
it wrote that tag in the yang file as well.

This patch will check if there is a } with
the <CODE ENDS> tag and if there is it will
remove this tag from that line

We are also adding new feature where user
can set strict-name option. This will
force the xym-tool tool to search for the
yang module name instead of using the
module name given after <CODE BEGINS> tag.
For example:
<CODE BEGINS> file "ONF_TR-512_CoreModel_Topolgy.yang"

// Contents of "Topology IM Draft for IETF-94"
module ietf-TopologyIM {

without strict-name we would receive file
called ONF_TR-512_CoreModel_Topolgy.yang
and with strict-name we would receive file
called ietf-TopologyIM.yang

Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>