Add support for extracting modules with their YANG Semver.
This adds two new arguments, --extract-semver and --semver-symlink,
which default to False. . The first argument will trigger the module to
be extracted (in addition to its @<revision> filename) a #<semver>
filename, which is a copy of the canonical @<revision> file.  The latter
argument turns that copy into a symlink.

This is being done to support the YANG module versioning drafts
currently in the IESG's hands.